src/main/java/com/zalinius/wanikaniapi/data/subject/MeaningData.java
package com.zalinius.wanikaniapi.data.subject;

import com.fasterxml.jackson.annotation.JsonIgnoreProperties;

@JsonIgnoreProperties(ignoreUnknown = true)
public class MeaningData {
	
	private String meaning;
	private boolean primary;
	
	public MeaningData() {
	}

	public String getMeaning() {
		return meaning;
	}

	public void setMeaning(String meaning) {
		this.meaning = meaning;
	}

	public boolean isPrimary() {
		return primary;
	}

	public void setPrimary(boolean primary) {
		this.primary = primary;
	}
	
	

}
